+++ src/main/java/net/silentchaos512/gear/data/part/PartsProvider.java
@@ -83,6 +83,14 @@ public class PartsProvider implements IDataProvider {
                 .trait(Const.Traits.ROAD_MAKER, 1)
                 .display(GearType.EXCAVATOR, PartType.MISC_UPGRADE, new MaterialLayer(SilentGear.getId("road_maker"), Color.VALUE_WHITE))
         );
+        ret.add(upgradePart("misc/wide_plate", CraftingItems.WIDE_PLATE_UPGRADE)
+                .upgradeGearTypes(new GearTypeMatcher(false, GearType.HAMMER, GearType.EXCAVATOR))
+                .stat(ItemStats.DURABILITY, 0.1f, StatInstance.Operation.MUL1)
+                .stat(ItemStats.RARITY, 10, StatInstance.Operation.ADD)
+                .trait(Const.Traits.WIDEN, 1)
+                .display(GearType.HAMMER, PartType.MISC_UPGRADE, new MaterialLayer(SilentGear.getId("wide_plate"), Color.VALUE_WHITE))
+                .display(GearType.EXCAVATOR, PartType.MISC_UPGRADE, new MaterialLayer(SilentGear.getId("wide_plate"), Color.VALUE_WHITE))
+        );
         ret.add(upgradePart("misc/red_card", CraftingItems.RED_CARD_UPGRADE)
                 .upgradeGearTypes(GearTypeMatcher.ALL)
                 .stat(ItemStats.RARITY, -5, StatInstance.Operation.ADD)