+++ src/main/java/net/silentchaos512/gear/item/gear/CoreHammer.java
@@ -9,12 +9,12 @@ import net.minecraft.util.math.RayTraceResult;
 import net.minecraft.world.World;
 import net.minecraftforge.common.ToolType;
 import net.silentchaos512.gear.api.item.GearType;
-import net.silentchaos512.gear.util.IAOETool;
+import net.silentchaos512.gear.util.IAoeTool;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 
-public class CoreHammer extends CorePickaxe implements IAOETool {
+public class CoreHammer extends CorePickaxe implements IAoeTool {
     public CoreHammer() {
         super(ImmutableSet.of());
     }
@@ -26,7 +26,7 @@ public class CoreHammer extends CorePickaxe implements IAOETool {
 
     @Nonnull
     @Override
-    public ToolType getAOEToolClass() {
+    public ToolType getAoeToolType() {
         return ToolType.PICKAXE;
     }
 
@@ -38,6 +38,6 @@ public class CoreHammer extends CorePickaxe implements IAOETool {
 
     @Override
     public boolean onBlockStartBreak(ItemStack itemstack, BlockPos pos, PlayerEntity player) {
-        return IAOETool.BreakHandler.onBlockStartBreak(itemstack, pos, player);
+        return IAoeTool.BreakHandler.onBlockStartBreak(itemstack, pos, player);
     }
 }