+++ CHANGELOG.md
@@ -3,8 +3,10 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [2.6.12] - 2021-03-20
 ### Added
+- Widen trait. Increases the area-of-effect for hammers and excavators. Max level 3, only 1 is obtainable by default.
+- Wide plate upgrade, which gives the Widen trait (level 1)
 - Config option to disable all conversion recipes
 ### Changed
 - Small change to durability and repair efficiency of shears